# internal/dto/tokens.go
package dto

type IssueTokensResponse struct {
	AccessToken  string `json:"access_token"`
	RefreshToken string `json:"refresh_token"`
}

type RefreshTokensRequest struct {
	AccessToken string `json:"access_token"`
	RefreshToken string `json:"refresh_token"`
}

type RefreshTokensResponse struct {
	AccessToken string `json:"access_token"`
	RefreshToken string `json:"refresh_token"`
}

type GetCurrentUserIDResponse struct{
	UserID string `json:"user_id"`
}


type FailureResponse struct{
	Error string `json:"error"`
}